+++ examples/demo-styles/index.scss
@@ -41,6 +41,6 @@
 @import "./upload.scss";
 @import "./divider.scss";
 @import "./image.scss";
-@import "./infiniteScroll.scss";
+@import "./infinite-scroll.scss";
 @import "./avatar.scss";
